ТЗ11 - 4 итерация

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -52,8 +52,8 @@ public class UserController {
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getCommonFriends(@PathVariable Integer id,
-                                       @PathVariable Integer otherId) {
+    public List<User> getCommonFriends(@PathVariable long id,
+                                       @PathVariable long otherId) {
         return userService.getCommonFriends(id, otherId);
     }
